+++ Test/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>moveTest</name>
+	<name>Test</name>
 	<comment></comment>
 	<projects>
 	</projects>
