+++ Test/src/test/Test.java
@@ -1,4 +1,4 @@
-package moveTest;
+package test;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
